Explicit cast to int to avoid KeyErrors #22 (#23)
If null values are sampled in one time step (e.g. nodata classes or masked areas), the resulting column will be cast to float before the nans can be dropped. This can result in a KeyError when max classes are filtered. Explicitly casting to int after dropping nans solves the issue.