Empty collection testing (#53)
* Swap black for ruff format

* Redefine ruff line length

* Modify error handling for empty collections

It looks like Earth Engine no longer raises a specific error for
sampling random points within a single point, so the error handling
and testing were modified for the more specific case of an empty
sampling region, e.g. a FeatureCollection with zero features.

In the process, the sampling error handling was refactored and a
new SamplingError class was added. SamplingError inherits from
ValueError which was the previous error type, so this shouldn't
affect any downstream error handling.

* Fix pandas deprecation warning

* Update auth mode

* Update auth mode

* Reset auth mode

* Update CI