Fix bug when all classes are interactively hidden
Hiding all classes would throw an exception related to mismatched
data types during some internal update of the widget state. I traced
that back to the `SankeyPlot.plot.data[0].link` update and found it
broke if you used the link data from a `go.Sankey` that was
initialized with `link_kwargs`. Instead, intitializing with an empty
`go.Sankey` seems to solve the issue. That only arises if all classes
are hidden, so I just threw in an early return of an empty `go.Sankey`
in that case.