Only validate dataset years after exceptions
Pre-validating the years prevented users accessing newly added
assets without a sankee update. Instead, we now try to get the
requested years and let GEE fail if they're unavailable. This is
in lieu of a more direct server-side check for the available years,
which would 1) add latency and 2) be complicated with collections
that aren't stored by year.