Catch missing years for LCMS and CCAP (#62)
The previous implementation of throwing ImageUnavailableError relied
on too specific of an EEException message that won't necessarily be
raised by Dataset subclasses, including LCMS. It's safer to catch any
exception during sankify and validate years then.

This also separates the fetching of a given year image from the
processing of that image so that Dataset subclasses don't need to
duplicate processing, like setting visualization properties.
Previously that wasn't being done consistently for LCMS and CCAP.