Add GLC FCS30D datasets to premade datasets #52
Add a subclass to the Datasets class:

Changes from the superclass
-------------------------
GLC FCS30D uses 2 separate ee.ImageCollections. Each collection contains 2 images. Each image contains 1 band for each year of the dataset. It contains 1 band for each year's data. Getting the data for a given year means selecting the appropriate band.

1. collection property
This function mosaics them, renames the bands to the corresponding years and combines them into a new image collection with 1 image for five-year bands and 1 for annual bands.

2. _fetch_year_image method
Flattening the ee.ImageCollection collection to a single image with ee.ImageCollection.toBands() renames the bands with to [image index]_[band name]. The five-year dataset is index 0 and annual is index 1. The band is renamed to match the band property.

3. _list_years method
The list of years in GLC FCS30D is equivelant to the list of bands. Get the list of bands and convert to ee.Numbers.

Added a test for get_year function of the subclass
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   sankee/datasets.py
	modified:   tests/test_datasets.py